% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetLogData.R
\name{GetLogData}
\alias{GetLogData}
\title{Pull log data for a given log from a WITSML Store}
\usage{
GetLogData(
  url = NULL,
  user = NULL,
  password = NULL,
  uidWell = "us_27990298",
  uid = "us_27990298_wb1_log_dfr_time_1s",
  curves = c("TIME", "DEPT", "BDEP"),
  startDateTimeIndex = NULL,
  endDateTimeIndex = NULL,
  startIndex = NULL,
  endIndex = NULL,
  printProgress = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{url}{url for the WITSML Data Store}

\item{user}{Username for the authenticating user}

\item{password}{Username for the authenticating user}

\item{uidWell}{Unique Identifier for the Parent Well}

\item{uid}{Unique Identifier for the Log}

\item{curves}{Character vector containing a list of mnemonic keys for log results required}

\item{startDateTimeIndex}{An optional UTC character timestamp specifying the desired start of a time based log}

\item{endDateTimeIndex}{An optional UTC character timestamp specifying the desired end of a time based log}

\item{startIndex}{An optional depth value specifying the desired start of a depth based log}

\item{endIndex}{An optional depth value specifying the desired end of a depth based log}

\item{printProgress}{Option to print the end of each data chunk received when pagination is required}
}
\value{
A data.table object with the specified log data
}
\description{
Pull log data for a given log from a WITSML Store
}
\examples{
RWITSML::GetLogData(url = "https://hub.us.pason.com/hub/witsml/store",
                    user = hpidc::GetKeys()$pason$user,
                    password = hpidc::GetKeys()$pason$password,
                    uidWell = "us_27990298",
                    uid = "us_27990298_wb1_log_dfr_time_1s",
                    curves = c("TIME","DEPT","BDEP"),
                    startDateTimeIndex = "2022-04-01T00:00:00+00:00",
                    endDateTimeIndex = "2022-04-01T00:04:59+00:00",
                    verbose = FALSE)

RWITSML::GetLogData(url = "https://hub.us.pason.com/hub/witsml/store",
                    user = hpidc::GetKeys()$pason$user,
                    password = hpidc::GetKeys()$pason$password,
                    uidWell = "us_27990298",
                    uid = "us_27990298_wb1_log_1",
                    curves = c("GAM","INCL","AZ"),
                    startIndex = 1,
                    endIndex = 24645,
                    verbose = FALSE)

}
