% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetLogHeader.R
\name{GetLogHeader}
\alias{GetLogHeader}
\title{Pull a specific log header from a WITSML Store}
\usage{
GetLogHeader(
  url = NULL,
  user = NULL,
  password = NULL,
  uidWell = NULL,
  uid = NULL,
  verbose = FALSE
)
}
\arguments{
\item{url}{url for the WITSML Data Store}

\item{user}{Username for the authenticating user}

\item{password}{Username for the authenticating user}

\item{uidWell}{Unique Identifier for the Parent Well}

\item{uid}{Unique Identifier for the Log}
}
\value{
A nested list with all log headers for a given well
}
\description{
Pull a specific log header from a WITSML Store
}
\examples{
RWITSML::GetLogHeader(url = "https://hub.us.pason.com/hub/witsml/store",
                       user = hpidc::GetKeys()$pason$user,
                       password = hpidc::GetKeys()$pason$password,
                       uidWell = "us_27990298",
                       uid = "us_27990298_wb1_log_dfr_time_1s",
                       verbose = FALSE)

}
