% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetLogs.R
\name{GetLogs}
\alias{GetLogs}
\title{Pull a list of logs for a given well from a WITSML Store}
\usage{
GetLogs(
  url = NULL,
  user = NULL,
  password = NULL,
  uidWell = NULL,
  verbose = FALSE
)
}
\arguments{
\item{url}{url for the WITSML Data Store}

\item{user}{Username for the authenticating user}

\item{password}{Username for the authenticating user}

\item{uidWell}{Unique Identifier for the Parent Well}
}
\value{
A nested list with all log headers for a given well
}
\description{
Pull a list of logs for a given well from a WITSML Store
}
\examples{
RWITSML::GetLogs(url = "https://hub.us.pason.com/hub/witsml/store",
                 user = hpidc::GetKeys()$pason$user,
                 password = hpidc::GetKeys()$pason$password,
                 uidWell = "us_28147307",
                 verbose = FALSE)

}
