% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetTrajectories.R
\name{GetTrajectories}
\alias{GetTrajectories}
\title{Pull a list of trajectories for a given well from a WITSML Store}
\usage{
GetTrajectories(
  url = NULL,
  user = NULL,
  password = NULL,
  uidWell = NULL,
  verbose = FALSE
)
}
\arguments{
\item{url}{url for the WITSML Data Store}

\item{user}{Username for the authenticating user}

\item{password}{Username for the authenticating user}

\item{uidWell}{Unique Identifier for the Parent Well}
}
\value{
A nested list with all trajectories for a given well
}
\description{
Pull a list of trajectories for a given well from a WITSML Store
}
\examples{
RWITSML::GetTrajectories(url = "https://hub.us.pason.com/hub/witsml/store",
                         user = hpidc::GetKeys()$pason$user,
                         password = hpidc::GetKeys()$pason$password,
                         uidWell = "us_28147307"
                         verbose = FALSE)

}
