% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetWells.R
\name{GetWells}
\alias{GetWells}
\title{Pull a list of all wells from a WITSML Store}
\usage{
GetWells(url = NULL, user = NULL, password = NULL, verbose = FALSE)
}
\arguments{
\item{url}{url for the WITSML Data Store}

\item{user}{Username for the authenticating user}

\item{password}{Username for the authenticating user}

\item{verbose}{if TRUE, verbose output from underlying curl functions will be printed to console. Useful for troubleshooting}
}
\value{
A data.table with all wells available
}
\description{
Pull a list of all wells from a WITSML Store
}
\examples{
RWITSML::GetWells(url = "https://hub.us.pason.com/hub/witsml/store",
                  user = hpidc::GetKeys()$pason$user,
                  password = hpidc::GetKeys()$pason$password,
                  verbose = FALSE)

RWITSML::GetWells(url = "https://witsml.welldata.net/witsml/wmls.asmx",
                  user = hpidc::GetKeys()$totco$user,
                  password = hpidc::GetKeys()$totco$password,
                  verbose = FALSE)

}
