% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestHub.R
\name{TestHub}
\alias{TestHub}
\title{Test connection and capabilities of a WITSML Data Store}
\usage{
TestHub(url = NULL, user = NULL, password = NULL, verbose = TRUE)
}
\arguments{
\item{url}{url for the WITSML Data Store}

\item{user}{Username for the authenticating user}

\item{password}{Username for the authenticating user}
}
\value{
A summary of server capabilities
}
\description{
Test connection and capabilities of a WITSML Data Store
}
\examples{
RWITSML::TestHub(url = "https://hub.us.pason.com/hub/witsml/store",
                 user = hpidc::GetKeys()$pason$user,
                 password = hpidc::GetKeys()$pason$password,
                 verbose = TRUE)

}
