% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetLogHeader.R
\name{GetLogHeader}
\alias{GetLogHeader}
\title{Pull a specific log header from a WITSML Store}
\usage{
GetLogHeader(
  url = NULL,
  user = NULL,
  password = NULL,
  uidWell = NULL,
  uidWellbore = NULL,
  uid = NULL,
  verbose = FALSE
)
}
\arguments{
\item{url}{url for the WITSML Data Store}

\item{user}{Username for the authenticating user}

\item{password}{Username for the authenticating user}

\item{uidWell}{Unique Identifier for the Parent Well}

\item{uidWellbore}{Unique Identifier for the Associated Wellbore. Not required for Pason, required for MD-Totco}

\item{uid}{Unique Identifier for the Log}

\item{verbose}{if TRUE, verbose output from underlying curl functions will be printed to console. Useful for troubleshooting}
}
\value{
A nested list with the log header for a specific log
}
\description{
Pull a specific log header from a WITSML Store
}
\examples{
RWITSML::GetLogHeader(url = "https://hub.us.pason.com/hub/witsml/store",
                       user = hpidc::GetKeys()$pason$user,
                       password = hpidc::GetKeys()$pason$password,
                       uidWell = "us_27990298",
                       uid = "us_27990298_wb1_log_dfr_time_1s",
                       verbose = FALSE)

RWITSML::GetLogHeader(url = "https://witsml.welldata.net/witsml/wmls.asmx",
                      user = hpidc::GetKeys()$totco$user,
                      password = hpidc::GetKeys()$totco$password,
                      uidWell = "9ce20147-e1d4-45db-a8cb-bedb3e50b65e",
                      uidWellbore = "9ce20147-e1d4-45db-a8cb-bedb3e50b65e",
                      uid = "Depth1",
                      verbose = TRUE)
}
