% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetWellBores.R
\name{GetWellBores}
\alias{GetWellBores}
\title{Pull a list of wellbores for a given well from a WITSML Store}
\usage{
GetWellBores(
  url = NULL,
  user = NULL,
  password = NULL,
  uidWell = NULL,
  verbose = FALSE
)
}
\arguments{
\item{url}{url for the WITSML Data Store}

\item{user}{Username for the authenticating user}

\item{password}{Username for the authenticating user}

\item{uidWell}{Unique Identifier for the Parent Well}

\item{verbose}{if TRUE, verbose output from underlying curl functions will be printed to console. Useful for troubleshooting}
}
\value{
A nested list with all wellbores for a given well
}
\description{
Pull a list of wellbores for a given well from a WITSML Store
}
\examples{
RWITSML::GetWellBores(url = "https://hub.us.pason.com/hub/witsml/store",
                      user = hpidc::GetKeys()$pason$user,
                      password = hpidc::GetKeys()$pason$password,
                      uidWell = "us_28147307",
                      verbose = FALSE)

RWITSML::GetWellBores(url = "https://witsml.welldata.net/witsml/wmls.asmx",
                      user = hpidc::GetKeys()$totco$user,
                      password = hpidc::GetKeys()$totco$password,
                      uidWell = "9ce20147-e1d4-45db-a8cb-bedb3e50b65e",
                      verbose = FALSE)
}
